+++ src/components/buttons/types/ButtonSize.ts
@@ -1,5 +1,3 @@
 export const buttonSizes = ["compact", "medium", "large"];
 
 export type ButtonSize = (typeof buttonSizes)[number];
-
-export const buttonSizeDefault: ButtonSize = "medium";