+++ src/components/widgets/shareButtons/ShareButtons.svelte
@@ -1,8 +1,17 @@
 <svelte:options customElement="khao-share-buttons" />
 
 <script lang="ts">
+  import { trapFocus } from "trap-focus-svelte";
   import Button from "../../buttons/button/Button.svelte";
-  import openInNewWindow from "../../../common/utils/openInNewWindow";
+
+  function openInNewWindow(
+    url: string,
+    width: number = 700,
+    height: number = 650
+  ) {
+    const windowFeatures = `left=100,top=100,width=${width},height=${height}}`;
+    window.open(url, "new", windowFeatures);
+  }
 
   function isAbleToPrint(): boolean {
     return window.print !== undefined;
@@ -52,7 +61,7 @@
   export let pinterestButtonTitle: string | null = pinterestButtonLabel;
 </script>
 
-<div class="share-buttons">
+<div class="share-buttons" use:trapFocus>
   {#if showPrintButton === "true" && isAbleToPrint()}
     <Button
       label={printButtonLabel || "No Label"}